Updated algorithm
Reactions take 2 or 3 inputs. If there are 2 inputs, either 1 or 2 random outputs (neither of which are inputs) are produced.
If there are 3 inputs, either 2 or 3 outputs are produced. One of these is the reagant, and none of the others are any of the inputs.
Has far fewer edges than even a scale-free graph. Distribution is binomial and falls off with 1/(k! * (n-k)!) for a compounds participated in k out of n reactions.